temperature streaming to the client works now for the develop device

+++ PlantMonitorControl/Features/ImageTaking/FileStreamingReader.cs
@@ -5,7 +5,7 @@ namespace PlantMonitorControl.Features.ImageTaking;
 public interface IFileStreamingReader
 {
     Task<FileInfo> ReadNextFile(string imagePath, int counter, CameraTypeInfo cameraInfo, CancellationToken token);
-    Task<FileInfo> ReadFromFile(CameraTypeInfo cameraInfo, string currentPath, CancellationToken token);
+    Task<FileInfo> ReadFromFile(CameraTypeInfo cameraInfo, string currentPath, CancellationToken token, bool deleteFile = true);
     Task<FileInfo> ReadNextFileWithSkipping(string imagePath, int counter, int howManyMoreRecentImagesMayExist, CameraTypeInfo cameraInfo, CancellationToken token);
 }
 
@@ -23,7 +23,7 @@ public record struct FileInfo(DateTime CreationDate, int NewCounter, byte[]? Fil
     }
 };
 
-public class FileStreamingReader : IFileStreamingReader
+public class FileStreamingReader(ILogger<FileStreamingReader> logger) : IFileStreamingReader
 {
     public const string CounterFormat = "000000";
     private static readonly string s_irEnding = CameraType.IR.Attribute<CameraTypeInfo>().FileEnding;
@@ -43,26 +43,41 @@ public class FileStreamingReader : IFileStreamingReader
         return result;
     }
 
-    public async Task<FileInfo> ReadFromFile(CameraTypeInfo cameraInfo, string currentPath, CancellationToken token)
+    public async Task<FileInfo> ReadFromFile(CameraTypeInfo cameraInfo, string currentPath, CancellationToken token, bool deleteFile = true)
     {
-        int temperatureInK = default;
-        var bytesToSend = cameraInfo.FileEnding == s_irEnding ? currentPath.GetBytesFromIrFilePath(out temperatureInK) : await File.ReadAllBytesAsync(currentPath, token);
-        var creationTime = File.GetCreationTimeUtc(currentPath);
-        File.Delete(currentPath);
-        return new(creationTime, 0, bytesToSend, temperatureInK);
+        try
+        {
+            int temperatureInK = default;
+            var bytesToSend = cameraInfo.FileEnding == s_irEnding ? currentPath.GetBytesFromIrFilePath(out temperatureInK) : await File.ReadAllBytesAsync(currentPath, token);
+            var creationTime = File.GetCreationTimeUtc(currentPath);
+            if (deleteFile) File.Delete(currentPath);
+            return new(creationTime, 0, bytesToSend, temperatureInK);
+        }
+        catch (Exception ex)
+        {
+            logger.LogError("Could not read from file: {error}\n{stacktrace}", ex.Message, ex.StackTrace);
+            return default;
+        }
     }
 
-    private static int SkipFiles(string imagePath, int counter, int skipCounter, CameraTypeInfo cameraInfo)
+    private int SkipFiles(string imagePath, int counter, int skipCounter, CameraTypeInfo cameraInfo)
     {
         if (Directory.GetFiles(imagePath, $"{(counter + skipCounter).ToString(CounterFormat)}*{cameraInfo.FileEnding}").Length != 0)
         {
-            for (var i = 0; i < skipCounter; i++)
+            try
             {
-                var files = Directory.GetFiles(imagePath, $"{(counter + skipCounter)
-                    .ToString(CounterFormat)}*{cameraInfo.FileEnding}");
-                foreach (var file in files) File.Delete(file);
+                for (var i = 0; i < skipCounter; i++)
+                {
+                    var files = Directory.GetFiles(imagePath, $"{(counter + skipCounter)
+                        .ToString(CounterFormat)}*{cameraInfo.FileEnding}");
+                    foreach (var file in files) File.Delete(file);
+                }
+                counter += skipCounter;
             }
-            counter += skipCounter;
+            catch (Exception ex)
+            {
+                logger.LogError("Filestream error: {error}\n{stacktrace}", ex.Message, ex.StackTrace);
+            }
         }
 
         return counter;
